+++ src/CompilerCxx/Naming.hs
@@ -98,7 +98,7 @@ mainSourceIncludes :: [String]
 mainSourceIncludes = ["#include \"logging.hpp\""]
 
 paramName :: ParamName -> String
-paramName p = "Param_" ++ tail (pnName p) -- Remove leading '#'.
+paramName p = "Param_" ++ tail (show p) -- Remove leading '#'.
 
 variableName :: VariableName -> String
 variableName v = "Var_" ++ show v