+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # Revision history for zeolite-lang
 
+## 0.14.0.0  -- ????-??-??
+
+### Language
+
+* **[breaking]** Adds the `#self` meta-type. This is an implicit type param that
+  is *always* the most-derived type. For example, if a `@value interface` has a
+  return type of `#self` then any implementation must return *its own type* from
+  that function. (This is a breaking change because it reserves the param name
+  `#self`.)
+
 ## 0.13.0.0  -- 2021-03-17
 
 ### Language